mmc: cmdq_hci: Do not handle other requests in case of error.
There can be case where controller may raise success for
rew queued requests while error for 1 of the requests.
In case of error we were handling the error but proceeding
and handling the requests which got completed by the controller.

As of now in cmdq-err handler mechanism, if any error occurs,
we re-queue all the requests back and request_queue.
So we need not call done for other requests which care completed
as this call lot of other contexts to run in parallel while error
handler is waiting to run.

Change-Id: I75214a870a13b39001d3d7a31e588c4c1113bed8
Signed-off-by: Ritesh Harjani <riteshh@codeaurora.org>
Signed-off-by: Sahitya Tummala <stummala@codeaurora.org>
Signed-off-by: Dusan Uveric <dusan.uveric9@gmail.com>