[EXPERIMENTAL] cpu_input_boost: Reduce frame boost timeout if vidc is active
The standard post-input timeout for frame boosting is 2500 ms. When
users constantly tap while watching a video, there can be a lot of
unnecessary CPU boosting. Dropping the timeout to 500 ms while videos
are playing dramatically reduces the impact of the issue while not
destroying performance in picture-in-picture scenarios where frame
boosting is more necessary than ever.

Ideally userspace could help us out, but it is hard to create a clean
way of passing powerhints to the kernel without requiring Magisk.

Signed-off-by: kdrag0n <dragon@khronodragon.com>